updated nutrition.handlebars and integrated fdcAPI.js into nutritionRoute.js
