+++ views/nutrition.handlebars
@@ -1,1 +1,7 @@
-<p>{{APIResult}}</p>
+<p>{{result.ingredientName}}</p>
+<p>Serving Size: {{result.servingSize}} {{result.unitOfMeasure}}</p>
+<p>Nutritional Facts:</p>
+{{#each result.nutrients as |nutrient|}}
+<p>{{nutrient.nutrient.name}}: {{nutrient.nutrient.number}} {{nutrient.nutrient.unitName}}</p>
+
+{{/each}}
